fix: half values into ingredients

+++ src/main/java/com/arthurcortez/javaproject/dto/CreateIngredientDto.java
@@ -1,4 +1,4 @@
 package com.arthurcortez.javaproject.dto;
 
-public record CreateIngredientDto(String name, String unityType, Integer unityValue) {
+public record CreateIngredientDto(String name, String unityType, Float unityValue) {
 }
